requirements.txt
Django
djangorestframework
django-cors-headers
pillow
django-filter
django-rest-auth
